DELTASPIKE-905 - move the code from ProxyClassGeneratorLookup to DeltaSpikeProxyFactory (1:1)
